requirements.txt
Flask
python-dotenv
langchain-core
langchain-groq
groq
Werkzeug
gunicorn
